Do not guess Brakeman's trademark owner
We assumed that "Brakeman" is now a trademarked by Synopsys, Inc
Presidentbeef said:
"A fair assumption, but it is not owned by Synopsys. You could argue
the trademark is owned by Brakeman, Inc. or by myself."
https://github.com/presidentbeef/brakeman/issues/1312

That same post also notes that Brakeman is under the
"Brakeman Public Use License".

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>